VOOK-003: Handler for `/videos` (#3)
## 🧑‍💻 Description
These changes contain the end-points for videos to perform CRUD
operations. They also include a new model called `TimeStamp` to
represent the duration of the video and other time stamps in the future,
it comes with MarshalJSON and UnmarshalJSON functions to perform binding
and rendering properly.

## ✅ Testing
I added unit testing for the majority of the functionality of the
end-points and confirm they are working via PostMan app. Also, included
several test cases for TimeStamp.

## 🖼️ Evidence
Verifying the container continue building locally and run with `docker
compose up --build`:

![image](https://github.com/zatarain/note-vook/assets/539783/8a0b5226-cebd-472d-ae1c-42dafc37f36e)

Played with the API using PostMan app, tested all the end-points:

![image](https://github.com/zatarain/note-vook/assets/539783/76aa4548-f792-4cfc-8bf8-e2a5815a27e8)

---------

Signed-off-by: Ulises Tirado Zatarain <ulises.tirado@cimat.mx>