VOOK-004: Annotation end-points (#4)
## 🧑‍💻 Description
These changes aim to add functionality for the end-point to manage the
annotations (add, edit, delete). In order to see the list of
annotations, you can see them using `GET /videos/:id`. They also are
delete in cascade when the associated vide is removed.

## ✅ Testing
I tested the new end-points manually in Postman and added more unit
testing for video related with the association video-annotation. I also
was following the queries in the command line logs.

## 🖼️ Evidence
Following is one screen shot of the testing in Postman:

![image](https://github.com/zatarain/note-vook/assets/539783/3b271659-7424-4664-bfbe-7819716a1b20)

I also was following the queries in the command line logs:

![image](https://github.com/zatarain/note-vook/assets/539783/30765881-a3fe-4d38-94fb-c2c0729062de)

---------

Signed-off-by: Ulises Tirado Zatarain <ulises.tirado@cimat.mx>